refactor(board-recognition): split moku-detector into class + postprocess
moku-detector.ts: 747 → 314 lines.

Pre/postprocess plumbing (sigmoid, logStoneStats, corner geometry
helpers, the 640×640 ImageNet preprocess, RT-DETR postprocess with
2/3/4-corner inference, and mapStonesToGrid) moved to
moku-postprocess.ts (442 lines, under the 500 max for pure logic).

moku-detector.ts now holds the MokuDetector class lifecycle only —
session init with optimization-level + freeDimensionOverrides
fallback chains, detect() with caching, fast refilter() path,
dispose().

Public exports (MokuDetector, mapStonesToGrid, expandCorners,
clearModelCache, DEFAULT_THRESHOLD, MokuDetectorConfig,
MokuDetectOptions, ProgressCallback) remain importable from
'@kaya/board-recognition' unchanged.

Verified by the package's 28 existing unit tests.
